Added up to 11 comment levels to be imported. Changed "private" to "draft" post status.
I had to migrate a full site with a lot of nodes and a lot of nested conversations in comments. Adding migration path for these extra levels allowed me to recover 100% of comments, where migrating only the first reply level recovered around 83% of comments. Also, changed "private" to "draft" when post is not public (unpublished Drupal nodes no longer appear as published in private mode, but they remain as draft).